move e2e tests to test package Remove test/* files from coverage
