+++ Makefile
@@ -27,10 +27,10 @@ test: mocks
 	go test -v ./...
 
 .PHONY: coverage
-coverage: 
-	go test ./... -coverprofile coverage.cov -coverpkg ./...
-	grep -v "_mock.go" coverage.cov > coverage_no_mocks.cov
-	mv coverage_no_mocks.cov coverage.cov
+coverage: mocks
+	go test ./... -coverprofile coverage.cov -coverpkg ./... || true
+	grep -v "_mock.go" coverage.cov > coverage_no_mocks.cov && mv coverage_no_mocks.cov coverage.cov
+	grep -v "test/" coverage.cov > coverage_no_mocks.cov && mv coverage_no_mocks.cov coverage.cov
 	go tool cover -func coverage.cov
 
 .PHONY: clean-mocks
